+++ Package/CMakeCPack.cmake
@@ -219,7 +219,7 @@ if(WIN32)
         Delete '$DESKTOP\\\\TransformCoordinate.lnk'
         ")
 endif()
-############### NSIS ###################
+############### NSIS end ###################
 
 # 设置传递给 CPack 的配置文件。主要作用是把 CMAKE_* 变量传递到 CPACK 中
 configure_file("${CMAKE_SOURCE_DIR}/Package/CMakeCPackOptions.cmake.in"